+++ project/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Ticket-0042 : refactor file uploader service.
 - Ticket-0043 : refactor login link service.
+- Ticket-0044 : refactor login link service again.
 
 ### Deprecated